+++ app.js
@@ -13,10 +13,9 @@ const router = express.Router()
 const routerIndex = require('./routes/index')
 const routerSobrenos = require('./routes/sobrenos')
 const routerCarrinho = require('./routes/carrinho')
-const routerCadastro = require('./routes/cadastro')
+
 const routerAreadocliente = require('./routes/areadocliente')
 const routerProdutos = require('./routes/produtos')
-
 
 
 app.use(express.urlencoded({extended: false}))
@@ -30,8 +29,7 @@ app.set('view engine', 'ejs')
 app.use('/', routerIndex)
 app.use('/sobrenos', routerSobrenos)
 app.use('/carrinho', routerCarrinho )
-app.use('/cadastro', routerCadastro)
-app.use('/areadoclientelogin', routerAreadocliente)
+app.use('/areadocliente', routerAreadocliente)
 //app.use('/areadoclientelogin/criar', routerAreadocliente)
 app.use('/produtos', routerProdutos)
