+++ controllers/areadoclienteControllers.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 const path = require('path');
-// const {validationResult} = require('express-validator')
+
 
 const usuariosFilePath = path.join(__dirname, '../data/cadastroClientes.json');
 const usuarios = JSON.parse(fs.readFileSync(usuariosFilePath, 'utf-8'));
@@ -11,24 +11,29 @@ const areadoclienteControllers = {
    },
 
     viewForm: (req, res) =>{
-        return res.render('areadoclientelogin')
+        return res.render('cadastro')
     },
     
     salvarForm: (req, res) =>{
-        let errors = validationResult(req)
-        if (!errors.isEmpty()){
-            return res.render('areadoclientelogin', {errors: errors.errors})
-        }
+        // let errors = validationResult(req)
+        // if (!errors.isEmpty()){
+        //     return res.render('areadoclientelogin', {errors: errors.errors})
+        // }
+        console.log(req.body)
 
-        let usuario = {
+        let novoUsuario = {
             id: usuarios[usuarios.length -1].id + 1,
-            usuarios: usuario,
-            // req.body,
+            nome: req.body.nomeCadastro,
+            email: req.body.emailCadastro,
+            password: req.body.passwordCadastro,
+
         }
-        usuario.push(novoUsuario)
+        usuarios.push(novoUsuario)
         fs.writeFileSync(usuariosFilePath, JSON.stringify(usuarios, null, ' '))
         res.redirect('/')
     },
+
+    
 }
