+++ controllers/produtosControllers.js
@@ -17,8 +17,30 @@ const produtosControllers = {
             res.render('produtos', { produtos: resultado })
 
         }
-        
 
+    },
+
+    //Edit
+    edit: (req, res) => {
+    let id = req.params.id
+    let productToEdit = products.find(product => product.id == id)
+    res.render('pageProductEdit', {productToEdit})
+    },
+    //Update POST
+    update: (req, res) => {
+    let id = req.params.id;
+    let productToEdit = products.find(product => product.id == id)
+    let image;
+    if (req.files[0] != undefined){
+        image = req.files [0].filename
+    }else{
+        image = productToEdit.image
     }
+    productToEdit = {
+        id: productToEdit.id,
+        ...req.body,
+        image: image,
+    }}
+
 }
 module.exports = produtosControllers;
