+++ data/cadastroClientes.json
@@ -1,0 +1,20 @@
+[
+ {
+  "id": 1,
+  "nome": "Diogenes",
+  "email": "dio@gmail.com",
+  "password": "1234"
+ },
+ {
+  "id": 2,
+  "nome": "luciana",
+  "email": "COLCHEI@HOTMAIL.COM",
+  "password": "c vzvc"
+ },
+ {
+  "id": 3,
+  "nome": "luciana",
+  "email": "COLCHEI@HOTMAIL.COM",
+  "password": " gnxfn"
+ }
+]
