+++ middlewares/verifyUserLogged.js
@@ -1,9 +1,11 @@
+
+
 function admin(req, res, next){
     let user = req.session.userLogged;
         if (user){
                 next();
-                res.redirect('/')
-            }else{
+               
+        }else{
                 const error = new Error ('Usuario nao tem privilegios')
         error.status = "401"
         throw error;
