+++ package.json
@@ -21,6 +21,7 @@
     "ejs": "^3.1.8",
     "express": "^4.18.1",
     "express-session": "^1.17.3",
-    "express-validator": "^6.14.2"
+    "express-validator": "^6.14.2",
+    "method-override": "^3.0.0"
   }
 }
