--- public/data/cadastroClientes.json
@@ -1,17 +1,0 @@
-[
-    {
-     "id": 1,
-     "username": "Diogenes",
-     "password": "123344"  
-    },
-    {
-     "id": 2,
-     "username": "Jonatas",
-     "password": "004441"
-    },
-    {
-       "id": 3,
-       "username": "Manuela",
-       "password": "458871"
-    }
-   ]
