--- routes/cadastro.js
@@ -1,10 +1,0 @@
-const express = require('express')
-const router = express.Router()
-
-const cadastroControllers = require('../controllers/cadastroControllers')
-router.get('/', cadastroControllers.index)
-
-router.get('/criar', cadastroControllers.viewForm)
-router.post('/criar', cadastroControllers.salvarForm)
-
-module.exports = router
