--- routes/formEdit.js
@@ -1,3 +1,0 @@
- 
-router.get('/edit/:id', formEditController.edit); 
-router.put('/edit/:id', upload.any(), formEditController.update);
