+++ views/areadoclienteLogin.ejs
@@ -9,17 +9,17 @@
     <section calss="login-content">
             <h2 id="login-title">Faça seu Login</h2>
             <div class="formulario-login">
-                <form class="formulario-login-content" action="/areadoclientelogin/criar" method="POST">
+                <form class="formulario-login-content" action="/areadocliente/login" method="POST">
 
                     <div class="datas-container">
                         <label class="datas-container" id="email" for="email">E-mail</label>
-                        <input class="datas-container" type="email" name="emailCriar" id="email" placeholder="E-mail" required>
+                        <input class="datas-container" type="email" name="email" id="email" placeholder="E-mail" required>
                         <label class="datas-container" id="senha" for="password">Senha</label>
-                        <input class="datas-container" type="password" name="passwordCriar" id="password" placeholder="Senha" required>
+                        <input class="datas-container" type="password" name="password" id="password" placeholder="Senha" required>
                         <input class="botao-entrar" id="botao-entrar" type="submit" name="Entrar" value="Entrar">
                     </div>
                         <span class="novo-cliente">Novo cliente?</span>
-                        <a class="cadastre-se" href= "/cadastro">Cadastre-se aqui →</a>    
+                        <a class="cadastre-se" href= "/areadocliente/cadastro">Cadastre-se aqui →</a>    
                 </form>
             </div>
   </main>
