+++ views/cadastro.ejs
@@ -9,7 +9,7 @@
     <section calss="login-content">
             <h2 id="login-title">Cadastro</h2>
             <div class="formulario-login">
-                <form class="formulario-login-content" action="/cadastro/criar" method="POST">
+                <form class="formulario-login-content" action="/areadocliente/criar" method="POST">
                     <span class="preencha-text">Preencha os campos abaixo para criar sua conta.</span>
                     <br>
 
@@ -21,7 +21,7 @@
                         <label class="datas-container" id="senha" for="password">Criar Senha*</label>
                         <input class="datas-container" type="password" name="passwordCadastro" id="password" required>
                         <label class="datas-container" id="senha" for="password">Confirme a Senha*</label>
-                        <input class="datas-container" type="password" name="passwordCadastro" id="password" required>
+                        <input class="datas-container" type="password" name="passwordCadastroConfirma" id="password" required>
                     </div>
 
                     <div class="newsletter-container">
