+++ views/partials/navbar.ejs
@@ -27,7 +27,7 @@
           </ul>
         </li>
         <li class="nav-item px-5 fs-5">
-          <a class="nav-link" href="/areadoclienteLogin">Área do Cliente</a>
+          <a class="nav-link" href="/areadocliente">Área do Cliente</a>
         </li>
         <li class="nav-item">
         </li>
